Set special categories on V8 JIT functions.
This is used with perf.data files which were created by first recording d8 (the V8 shell)
with perf record, and then processing the perf.data file with perf inject --jit.
V8 adds special prefixes to the symbol name to indicate the JIT tier.
With different categories, each of which has a different color, the different tiers
will be visible in the activity graph.